updated HC

+++ modules/mc/hc/hc_ggrd_auto_proto.h
@@ -34,4 +34,4 @@ void ggrd_indexx(int, double *, int *);
 float ggrd_gt_rms(float *, int);
 float ggrd_gt_mean(float *, int);
 /* sh_exp_ggrd.c */
-void sh_read_spatial_data_from_grd(struct sh_lms *, struct ggrd_gt *, unsigned short, int, long double *, long double *);
+void sh_read_spatial_data_from_grd(struct sh_lms *, struct ggrd_gt *, unsigned short, int, double *, double *);
